Change structure and Add backend directory
