+++ backend/users_api/urls.py
@@ -1,0 +1,11 @@
+from django.urls import path
+from . import views
+from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
+
+urlpatterns = [
+    path("", views.home, name="home"),
+    path("route/", views.route, name="route"),
+    path("token/", TokenObtainPairView.as_view(), name="get_token"),
+    path("token/refresh/", TokenRefreshView.as_view(), name="refresh_token"),
+    # path("user/register/", views.CreateUserView.as_view(), name="register"),
+]
