--- eightsteps/eightstepsmain/static/css/mainpage.css
@@ -1,361 +1,0 @@
-/* General Styles ----------------------------------------- */ 
-
-* {
-	margin: 0;
-	padding: 0;
-	box-sizing: border-box;
-}
-
-html,body {
-	width: 100%;
-	height: 100vh;
-	box-sizing: border-box;
-	background: rgba(24, 24, 24, 0.98);
-	color:aliceblue;
-	overflow:hidden;
-}
-
-.container-items {
-	position: absolute;
-	bottom: 0;
-	width: 100%;
-	height: 300px;
-	display: flex;
-	align-items: center;
-	cursor: pointer;
-}
-
-@media (max-width: 900px) {
-	.container-items {
-		height: 100px;
-	}
-	.item {
-		font-size: 20px;
-	}
-	.item:hover {
-		flex-grow: 1;
-		font-size: 20px;
-	}
-	.container-img-wrap {
-		display: none;
-	}
-	.navbar {
-		display: none !important;
-	}
-}
-
-.site-logo {
-    position: fixed;
-    left: 50%;
-    transform: translateX(-50%);
-    font-family: 'Oswald', sans-serif;
-    font-size: 36px;
-    line-height: 100px;
-    text-transform: uppercase;
-    display: flex;
-    align-items: center; /* Alinea verticalmente el icono con el texto */
-}
-
-.logo-icon {
-    width: 62px; /* Ajusta el tamaño del SVG según sea necesario */
-    height: auto;
-    margin-right: 3px; /* Elimina el margen entre el SVG y el texto */
-    vertical-align: middle; /* Alinea el SVG al centro verticalmente con el texto */
-    filter: invert(100%) sepia(0%) saturate(0%) hue-rotate(0deg) brightness(100%) contrast(100%);
-}
-
-.bold-text {
-    font-weight: 700; /* Negrita para "8STE" */
-    display: inline-block;
-}
-
-.normal-text {
-    font-weight: 400; /* Peso normal para "PS" */
-    display: inline-block;
-}
-
-.navbar {
-    position: relative; /* Usa relative si no necesitas que se quede fijo */
-    width: 80%;
-    height: 100px;
-    /* padding: 0 40px; */
-    display: flex;
-    justify-content: space-between;
-    font-family: oswald, Helvetica, sans-serif;
-    text-transform: uppercase;
-    line-height: 100px;
-}
-
-.site-menu {
-	display: flex;
-}
-
-.menu-item {
-	margin-left: 60px;
-}
-
-.header {
-    position: relative;
-    width: 80%;
-    display: flex;
-    justify-content: space-between;
-    top: 20%;
-    padding-top: 100px;
-    text-transform: uppercase;
-    font-family: "Roboto";
-    font-size: 40px;
-    z-index: 2;
-}
-
-.header-left span {
-	font-family: Oswald, Helvetica, sans-serif;
-	font-weight: bold;
-	font-size: 50px;
-}
-
-.link {
-	margin: 40px 0;
-}
-
-.link a {
-	font-family: Arial, Helvetica, sans-serif;
-	font-size: 16px;
-	color: #000;
-}
-
-.header-right {
-    position: relative;
-    cursor: pointer;
-}
-
-.hover-text {
-    display: none;
-	text-transform: none;
-	font-size: small;
-    left: 0;
-    background-color: #292929;
-    padding: 10px 15px;
-    border: 1px solid #292929;
-    border-radius: 8px;
-    white-space: nowrap;
-    opacity: 0;
-    transform: translateY(-20px) scale(0.95); /* Comienza desplazado y reducido */
-    transition: opacity 0.6s ease, transform 0.6s ease;
-    box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1); /* Sombra suave para dar profundidad */
-}
-
-.header-right:hover .hover-text {
-    display: block;
-    opacity: 1;
-    transform: translateY(0) scale(1); /* Desplaza a su lugar y expande suavemente */
-}
-
-
-/* Estilos para la pantalla de carga -----------------------------------------------------*/
-#loading-screen {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: #333; /* Fondo oscuro */
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 9999;
-    transition: opacity 0.5s ease-out;
-}
-
-/* Estilo para el logo de carga */
-.loading-content {
-    text-align: center;
-}
-
-.loading-logo {
-    width: 150px;
-    animation: fadeIn 3s ease-out;
-	filter: invert(100%) sepia(0%) saturate(0%) hue-rotate(0deg) brightness(100%) contrast(100%);
-
-}
-
-/* Animación para el logo */
-@keyframes fadeIn {
-    from {
-        opacity: 0;
-    }
-    to {
-        opacity: 1;
-    }
-}
-
-
-/* Estilos para el video y la línea separadora --------------------------------------------------------*/
-
-#background-video {
-	width: 80%;
-    max-width: 100%;
-    height: auto;
-    border: 3px solid #292929;
-    border-radius: 10px;
-    background-color: black;
-    object-fit: cover;
-    margin-top: 90px;
-    height: 220px;
-	box-shadow:  0px 0px 22px -16px rgb(255, 255, 255)
-}
-
-.circle {
-	height: 24px;
-	width: 24px;
-	border-radius: 24px;
-	background-color: black;
-	position: fixed; 
-	top: 0;
-	left: 0;
-	pointer-events: none;
-	z-index: 99999999; /* so that it stays on top of all other elements */
-    transition: transform 0.2s ease;
-
-  }
-  
-.circle:hover {
-transform: scale(5);
-}
-/* Searchbar ----------------------------------------- */ 
-
-.search-bar {
-    position: relative; /* Cambia a relative para que se mueva con el contenido */
-    display: flex;
-    align-items: center;
-    background-color: #ffffff;
-	padding: 10px 20px;
-    border-radius: 50px;
-	box-shadow:  0px 0px 36px -16px rgb(255, 255, 255);
-    font-family: Arial;
-    margin-top: 20px; /* Ajusta el margen superior si es necesario */
-	width: 80%;
-}
-
-
-select {
-    background-color: #ffffff;
-    border: none;
-    border-radius: 50px;
-    padding: 8px 12px;
-    margin-right: 10px;
-    font-size: 16px;
-}
-
-.select:focus {
-    outline: none;
-}
-
-.search-fields {
-    display: flex;
-    align-items: center;
-	width: 100%;
-}
-
-.input-field {
-    background-color: #ffffff;
-    border: none;
-    border-radius: 50px;
-    padding: 8px 12px;
-    margin-right: 10px;
-    font-size: 16px;
-    flex: 1;
-}
-
-#input-field-soy {
-    display: none;
-}
-
-.input-field:focus {
-    outline: none;
-}
-
-#andText {
-    margin-right: 10px;
-    font-size: 16px;
-	font-family: Arial, Helvetica, sans-serif;
-	color:#000;
-	text-transform: none;
-	font-weight: 250;
-}
-
-.wrapper {
-    display: flex;
-    flex-direction: column; /* Organiza los elementos verticalmente */
-    align-items: center; /* Centra horizontalmente todos los elementos */
-    padding: 10px;
-    width: 100%;
-    position: relative; /* Asegúrate de que tenga un contexto de posición relativo */
-}
-
-/* Botón */
-
-button {
-	position: relative;
-	display:block;
-	height: 45px;
-	width: 20%;
-	margin: 10px 7px;
-	margin-top: 60px;
-	padding: 5px 5px;
-	font-weight: 700;
-	font-size: 15px;
-	letter-spacing: 2px;
-	color: #ffffff;
-	border: 2px #ffffff solid;
-	border-radius: 50px;
-	box-shadow:  0px 0px 36px -16px rgb(255, 255, 255);
-	text-transform: uppercase;
-	outline: 0;
-	overflow:hidden;
-	background: none;
-	z-index: 1;
-	cursor: pointer;
-	transition:         0.08s ease-in;
-	-o-transition:      0.08s ease-in;
-	-ms-transition:     0.08s ease-in;
-	-moz-transition:    0.08s ease-in;
-	-webkit-transition: 0.08s ease-in;
-  }
-  
-  .slide:after{
-	content:"Ver trayectoria";
-	position:absolute;
-	width:100%;
-	height:100%;
-	left:0;
-	text-align:center;
-	-webkit-transition: all 400ms cubic-bezier(0.680, -0.550, 0.265, 1.550); 
-  }
-  
-  .slide:before {
-	content:"Descúbrela! →";
-	height:100%;
-	width:100%;
-	position:absolute;
-	color:#ffffff;
-	left:-100%;
-	opacity: 0;
-	-webkit-transition: all 500ms cubic-bezier(0.680, -0.550, 0.265, 1.550); 
-  }
-  
-  .slide:hover{
-	background:#ffffff;
-  }
-  
-  .slide:hover:before{
-	left:0;
-	opacity:1;
-	color:#000000;
-  }
-  .slide:hover:after{
-	left:100%;
-	opacity:0;
-  }
-  
-
-
