--- eightsteps/eightstepsmain/static/css/stepspage.css
@@ -1,241 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-html,
-body {
-  width: 100%;
-  height: 100%;
-  background: rgba(24, 24, 24, 0.98);
-  overflow: hidden;
-  color: #ffffff;
-}
-
-main {
-  height: 100%;
-  width: 100vw;
-  display: flex;
-  overflow: hidden;
-}
-
-section {
-  width: 120vw;
-  font-size: 36px;
-  line-height: 1em;
-  background: rgba(24, 24, 24, 0.98);;
-  font-family: "Roboto";
-
-}
-
-.wrap {
-  display: flex;
-  height: 100%;
-  font-size: medium;
-}
-
-.block__title {
-  position: relative;
-  white-space: nowrap;
-  transform: rotate(-90deg) translate(-50%);
-  text-align: right;
-}
-
-.block__number,
-.block__title {
-  text-transform: uppercase;
-  font-size: 2rem;
-  line-height: 1.06rem;
-  color: #ffffff;
-  font-family: "Roboto";
-
-}
-
-.section {
-  height: 100%;
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  left: -200px;
-}
-
-.blocks {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  z-index: 100;
-  pointer-events: none;
-}
-
-.block {
-
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 4rem;
-  height: 100%;
-  background: rgba(24, 24, 24, 0.98);;
-  z-index: 100;
-  padding: 2rem 1.25rem;
-
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  justify-content: space-between;
-
-  pointer-events: auto;
-  cursor: pointer;
-  border-left: 0.2rem solid #ffffff;
-  box-shadow:  0px 0px 20px -16px rgb(255, 255, 255);
-
-  font-weight: 500;
-
-  
-}
-
-.block[data-block-section="2"].init {
-  left: 100%
-}
-
-.block[data-block-section="2"].fixed {
-  left: 4rem;
-  border-left: 0.2rem solid #ffffff;
-  border-right: 0.2rem solid #ffffff;
-}
-
-.block[data-block-section="3"].init {
-  left: calc(100vw);
-}
-
-.block[data-block-section="3"].fixed {
-  left: 4rem;
-  border-right: 0.2rem solid #ffffff;
-
-}
-
-.block[data-block-section="4"].init {
-  left: calc(100vw);
-}
-
-.block[data-block-section="4"].fixed {
-  left: 4rem;
-  border-right: 0.2rem solid #ffffff;
-
-}
-
-.block[data-block-section="5"].init {
-  left: calc(100vw);
-}
-
-.block[data-block-section="5"].fixed {
-  left: 4rem;
-  border-right: 0.2rem solid #ffffff;
-}
-
-.block[data-block-section="6"].init {
-  left: calc(100vw);
-}
-
-.block[data-block-section="6"].fixed {
-  left: 4rem;
-  border-right: 0.2rem solid #ffffff;
-}
-
-.block[data-block-section="7"].init {
-  left: calc(100vw);
-}
-
-.block[data-block-section="7"].fixed {
-  left: 4rem;
-  border-right: 0.2rem solid #ffffff;
-}
-
-.block[data-block-section="8"].init {
-  left: calc(100vw);
-}
-
-.block[data-block-section="8"].fixed {
-  left: 4rem;
-  border-right: 0.2rem solid #ffffff;
-}
-
-
-.block[data-block-section="9"].init {
-  left: calc(100%);
-}
-
-.block[data-block-section="9"].fixed {
-  left: 4rem;
-  border-right: 0.2rem solid #ffffff;
-}
-
-
-body {
-  margin: 0;
-  font-family: Arial, sans-serif;
-}
-
-.navbar {
-  width: 4rem;
-  height: 100vh;
-  background-color: #181818;
-  position: fixed;
-  right: 0;
-  top: 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 1rem 0;
-  z-index: 99699999;
-  box-shadow:  0px 0px 20px -16px rgb(255, 255, 255);
-
-
-}
-
-.navbar-logo {
-  margin-bottom: 2rem;
-  filter: invert(100%) sepia(0%) saturate(0%) hue-rotate(0deg) brightness(100%) contrast(100%);
-  cursor: pointer;
-  z-index: 9999999999999999;
-}
-
-
-.logo {
-  width: 3rem;
-  height: auto;
-
-}
-
-.navbar-menu {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-  width: 100%;
-  z-index: 0;
-}
-
-.navbar-menu li {
-  width: 100%;
-  text-align: center;
-}
-
-.navbar-menu li a {
-  display: block;
-  color: white;
-  padding: 1rem 0;
-  text-decoration: none;
-  font-size: 0.9rem;
-}
-
-.navbar-menu li a:hover {
-  background-color: #333;
-}
-
-.content {
-  margin-right: 4rem; /* Para evitar que el contenido se superponga a la navbar */
-  padding: 2rem;
-}
