--- pruebas/searchbar.css
@@ -1,59 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    margin: 0;
-    background-color: #f5f8fa;
-}
-
-.search-bar {
-    display: flex;
-    align-items: center;
-    background-color: #ffffff;
-    border-radius: 50px;
-    padding: 10px 20px;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-}
-
-select {
-    background-color: #ffffff;
-    border: none;
-    border-radius: 50px;
-    padding: 8px 12px;
-    margin-right: 10px;
-    font-size: 16px;
-}
-
-.select:focus {
-    outline: none;
-}
-
-.search-fields {
-    display: flex;
-    align-items: center;
-}
-
-.input-field {
-    background-color: #ffffff;
-    border: none;
-    border-radius: 50px;
-    padding: 8px 12px;
-    margin-right: 10px;
-    font-size: 16px;
-    flex: 1;
-}
-
-#input-field-soy {
-    display: none;
-}
-
-.input-field:focus {
-    outline: none;
-}
-
-#andText {
-    margin-right: 10px;
-    font-size: 16px;
-}
