--- pruebas/searchbar.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="searchbar.css">
-    <title>Searchbar</title>
-</head>
-<body>
-    <div class="search-bar">
-        <select id="optionSelector" onchange="updateSearchBar()">
-            <option value="soy">Soy</option>
-            <option value="quiero">Quiero trabajar como</option>
-        </select>
-        <div id="searchFields" class="search-fields">
-            <input type="text" placeholder="Profesión o estudio actual" class="input-field" > 
-            <input type="text" placeholder="Tu trabajo deseado, puedes añadir en donde!." class="input-field" id="input-field-soy">
-            <span id="andText"> y quiero trabajar como</span>
-            <input type="text" placeholder="Trabajo deseado, puedes añadir en donde!" class="input-field">
-        </div>
-    </div>
-
-    <script src="searchbar.js"></script>
-</body>
-</html>
